ADDED : Create Contact

+++ src/com/bridgelabz/AddressBookSystem.java
@@ -4,13 +4,11 @@ import java.util.Scanner;
 
 public class AddressBookSystem {
     public static Scanner scanner = new Scanner(System.in);
-
     public static void main(String[] args) {
         System.out.println("Welcome to Address Book System !");
         Contact contact = addContact();
         System.out.println(contact);
     }
-
     public static Contact addContact() {
         Contact contact = new Contact();
         System.out.println("Enter First name");
@@ -26,7 +24,7 @@ public class AddressBookSystem {
         System.out.println("Enter Zip code");
         contact.setZip(Integer.parseInt(scanner.next()));
         System.out.println("Enter Phone number");
-        contact.setPhoneNumber(Integer.parseInt(scanner.next()));
+        contact.setPhoneNumber((int) Long.parseLong(scanner.next()));
         System.out.println("Enter Email");
         contact.setEmail(scanner.next());
         return contact;
